Implement the file size limitation
 It uses %95 of the given file size limit as the criteria, which refers to the MPEG4Writer.cpp in libstagefright.

 Change the type of file size limit from int to long to accept size over 2.14G which is the maximum value of java integer.

Bug: 167481981
Test: ./gradlew camera:camera-video:connectedAndroidTest

+++ camera/camera-video/src/main/java/androidx/camera/video/FileDescriptorOutputOptions.java
@@ -49,7 +49,7 @@ public abstract class FileDescriptorOutputOptions extends OutputOptions {
      * Gets the limit for the file length in bytes.
      */
     @Override
-    public abstract int getFileSizeLimit();
+    public abstract long getFileSizeLimit();
 
     /**
      * Gets the file descriptor instance.
@@ -80,7 +80,7 @@ public abstract class FileDescriptorOutputOptions extends OutputOptions {
          * <p>If not set, defaults to {@link #FILE_SIZE_UNLIMITED}.
          */
         @NonNull
-        public abstract Builder setFileSizeLimit(int bytes);
+        public abstract Builder setFileSizeLimit(long bytes);
 
         /** Builds the FileDescriptorOutputOptions instance. */
         @NonNull
